refactor: upgrade react toastify to v9
