+++ package.json
@@ -64,7 +64,7 @@
         "prettier": "^3.1.1",
         "react-ga4": "^1.4.1",
         "react-mde": "^11.5.0",
-        "react-toastify": "^8.2.0",
+        "react-toastify": "9.1.3",
         "typescript": "5.5.4",
         "vite": "5.4.2",
         "vite-plugin-dts": "4.0.3",
