+++ src/Shared/Services/ToastManager/index.ts
@@ -1,3 +1,3 @@
-export { default as ToastManager } from './toastManager.service'
+export { default as ToastManager, ToastManagerContainer } from './toastManager.service'
 export { ToastVariantType } from './types'
 export type { ToastProps } from './types'
